+++ src/main/java/org/brutusin/jsonsrv/caching/CachingInfo.java
@@ -13,19 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.brutusin.jsonsrv;
-
-import org.brutusin.jsonsrv.caching.CachingInfo;
+package org.brutusin.jsonsrv.caching;
 
 /**
  *
  * @author Ignacio del Valle Alles idelvall@brutusin.org
  */
-public abstract class JsonAction<I, O> {
+public interface CachingInfo {
 
-    public CachingInfo getCachingInfo(I input) {
-        return null;
-    }
-
-    public abstract O execute(I input) throws Exception;
 }